No issue: Move TestResource.kt to /helpers
